pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8.0,<3.9"
numpy = "^1.22.2"
Flask = "^2.1.1"
pandas = "^1.4.2"
python-dotenv = "^0.20.0"
bs4 = "^0.0.1"
discord = "^1.7.3"
lxml = "^4.8.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"